pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend service for portfolio application"
authors = ["Anubhavgupta14 <anubhavgu2002@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = {extras = ["standard"], version = "^0.116.1"}
uvicorn = {extras = ["standard"], version = "^0.35.0"}
python-dotenv = "^1.1.1"
google-generativeai = "^0.8.5"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
pypdf = "^5.8.0"
pyyaml = "^6.0.2"
tavily-python = "^0.7.10"
boto3 = "^1.39.14"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
black = "^24.4.2"
flake8 = "^7.0.0"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
)/
'''

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"